F refactor (#20)
* Update systemd

* Add bastion

* Add Vault to bastion

* Remove bastion files

* Run build

* Remove RUN_BUILD

* Add RUN_BUILD

* Removing RUN_BUILD

* Test with Run Build

* Start services

* Do not start services in install

* Put env var flag file in different dir

* Change env var file location

* Trigger build

* Sudo cat

* sudo tee

* Fix Vault -dev mode

* Update tests

* Update scripts and Vagrantfiles

* Limiting certain tests to AWS for now

* Fix up HashiStack Vagrantfile

* Add dev flags and update tests

* Update ports

* Update CI, Packer templates, and Vagrantfiles

* Add hashistack folder and fix build

* Create Consul and Vault ent builds

* Copy AMI to other regions

* Build OSS

* Add conditional branch logic to publish ci function

* Update

* Update repo location

* Create download script

* Install git

* Fix dir

* Update download

* Add ca-certificates to install

* Add 127.0.0.1 to resolv.conf

* Update PRODUCT names

* Remove unnecessary config

* Update Vagrantfiles with new test

* Remove un-used Rakefile's and move example env

* Fix ci-functions

* Hardcode aws template in CI

* Remove extra functions

* Update templates

* Update services

* Revert consul removal

* Fix CI

* Remove logger and use echo

* Remove data dir cleanup

* Update copy command

* Add pki folder to Consul config dir

* Update Vault PKI dir for certs, change config folder perms

* Add pki folder to tools

* Change tls directory name

* Update config file permissions

* Fix NOMAD_ADDR

* Remove example config files

* Start Docker after install

* Just download systemd config

* Update script branch name

* Update systemd install scripts for Nomad and Vault

* Remove Consul, Vault, Nomad Vagrantfiles and update HashiStack

* Remove HashiStack Vagrantfile

* Fix permissions dir

* Add HashiStack Vagrantfile back

* Fix Consul AWS packer template

* Bump versions and refactor setup scripts

* Update ci-functions.sh with new CI logic

* Give sudo permissions

* Fix tests and add Vagrantfiles

* Remove readable/writable by

* Update AMI names

* Update VCS_NAME

* Fix Docker install

* Remove Docker install comment

* Fixing Docker install

* Adding a quick test to verify Docker is installed correctly

* Removing container-selinux before installing docker

* Fixing yum erase

* Rename java script

* Pinning Docker version

* Update setcap

* Add config dir to dev install

* Update install scripts

* Update install scripts

* Fix setcap path

* Update install scripts

* Update versions and remove deregister to get past auth failure

* Revert force deregister

* Don't source variables

* Bumping relase version to 0.1.0-beta1

* Add shell access for Consul user

* Give Consul sudo access to certain commands for remote execution

* Disabling Docker test for RHEL

* Build dev1

* Build enterprise

* Fix ent

* Temporarily copy to us-east-1

* Move Product ent download url into their own function

* Remove comments

* Build OSS

* Fix RUN_PUBLISH

* Fix CONSU?L_HTTP_ADDR

* Build Vault 0.10.0

* Vault 0.10 Enterprise bump

* Update resolv.conf

* Bump Nomad to 0.8.0

* Bump Nomad 0.8.0 Enterprise

* Build 0.1.0

* Build 0.1.0 enterprise

* Update region copies

* Set Vault's dev-listen-address

* Don't run build

* Add Consul profile script

* Update f-refactor to master

+++ ci-functions.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # This script includes a set of generic CI functions to test Packer Builds.
 prepare () {
   rm -rf /tmp/packer
@@ -11,18 +13,22 @@ prepare () {
 }
 
 validate () {
-  for PRODUCT in $*; do
-    echo "Reviewing ${PRODUCT} template ...             "
-    cd "${BUILDDIR}/${PRODUCT}"
-    if /tmp/packer validate ${PRODUCT}.json; then
+  for TEMPLATE in $*; do
+    echo "cd into ${TEMPLATE} directory"
+    cd ${BUILDDIR}/$(echo ${TEMPLATE} | sed 's/-.*//')
+    echo "Reviewing ${TEMPLATE}.json template..."
+
+    if /tmp/packer validate ${TEMPLATE}.json; then
       echo -e "\033[32m\033[1m[PASS]\033[0m"
     else
       echo -e "\033[31m\033[1m[FAIL]\033[0m"
       return 1
     fi
+
     cd -
   done
-  echo "Reviewing shell scripts ..."
+
+  echo "Reviewing shell scripts..."
   if find . -iname \*.sh -exec bash -n {} \; > /dev/null; then
     echo -e "\033[32m\033[1m[PASS]\033[0m"
   else
@@ -35,59 +41,34 @@ gpg_import () {
   echo ${PGP_SECRET_KEY} | base64 -d | gpg --import
 }
 
-
-build () {
-export CONSUL_RELEASE="${CONSUL_VERSION}"
-export NOMAD_RELEASE="${NOMAD_VERSION}"
-export VAULT_RELEASE="${VAULT_VERSION}"
-
-if [[ $1 =~ .*ent ]] ; then
-  PRODUCT=${1/-ent/}
-  echo "Building $PRODUCT Enterprise template ...       "
-  export CONSUL_VERSION="${CONSUL_VERSION}+ent"
-  export NOMAD_VERSION="${NOMAD_RELEASE}+ent"
-  export VAULT_VERSION="${VAULT_VERSION}+ent"
-  export CONSUL_ENT_URL=${_CONSUL_ENT_URL}
-  export VAULT_ENT_URL=${_VAULT_ENT_URL}
-  export NOMAD_ENT_URL=${_NOMAD_ENT_URL}
-  export DISTRIBUTION="ent"
-else
-  PRODUCT=$1  
-  echo "Building $PRODUCT OSS template ...       "
-  export CONSUL_ENT_URL=''
-  export VAULT_ENT_URL=''
-  export NOMAD_ENT_URL=''
-  export DISTRIBUTION="oss"
-fi  
-
-cd "${BUILDDIR}/${PRODUCT}"
-export TMPDIR="/tmp/${PRODUCT}-$((1 + RANDOM % 100))"
-mkdir -p $TMPDIR
-if /tmp/packer build ${PRODUCT}.json ; then
-  echo -e "\033[32m${PRODUCT} \033[1m[PASS]\033[0m"
-else
-  echo -e "\033[31m${PRODUCT} \033[1m[FAIL]\033[0m"
-  return 1
-fi
-cd -
-}
-
 presign_ent_url () {
   if [ $# -eq 0 ]; then
+    echo -e "\033[31m\033[1m[FAIL - no variables provided]\033[0m"
+    return 1
+  fi
+
+  if [ $1 -eq 0 ]; then
     echo -e "\033[31m\033[1m[FAIL - no product name provided]\033[0m"
     return 1
   fi
-  
+
+  if [ $2 -eq 0 ]; then
+    echo -e "\033[31m\033[1m[FAIL - no product version provided]\033[0m"
+    return 1
+  fi
+
   _AWS_SECRET_ACCESS_KEY=$(echo $AWS_SECRET_ACCESS_KEY_BINARY | base64 -d | gpg -d -)
   _AWS_ACCESS_KEY_ID=${AWS_ACCESS_KEY_ID_BINARY}
   _REGION="us-east-1"
+  _PRODUCT=$1
+  _VERSION=$2
 
-  if [ $1 = "consul" ]; then
-    S3_URL=s3://${S3BUCKET}/consul-enterprise/${CONSUL_VERSION}/consul-enterprise_${CONSUL_VERSION}+ent_linux_amd64.zip
-  elif [ $1 = "vault" ]; then
-    S3_URL=s3://${S3BUCKET}/vault/prem/${VAULT_VERSION}/vault-enterprise_${VAULT_VERSION}+prem_linux_amd64.zip
-  elif [ $1 = "nomad" ]; then
-    S3_URL=s3://${S3BUCKET}/nomad-enterprise/${NOMAD_VERSION}/nomad-enterprise_${NOMAD_VERSION}+ent_linux_amd64.zip
+  if [ ${_PRODUCT} = "consul" ]; then
+    _S3_URL=s3://${S3BUCKET}/consul-enterprise/${_VERSION}/consul-enterprise_${_VERSION}+ent_linux_amd64.zip
+  elif [ ${_PRODUCT} = "vault" ]; then
+    _S3_URL=s3://${S3BUCKET}/vault/prem/${_VERSION}/vault-enterprise_${_VERSION}+prem_linux_amd64.zip
+  elif [ ${_PRODUCT} = "nomad" ]; then
+    _S3_URL=s3://${S3BUCKET}/nomad-enterprise/${_VERSION}/nomad-enterprise_${_VERSION}+ent_linux_amd64.zip
   else
     echo -e "\033[31m\033[1m[FAIL - invalid product selection for S3 enterprise URL]\033[0m"
     return 1
@@ -97,14 +78,93 @@ presign_ent_url () {
     AWS_ACCESS_KEY_ID=${_AWS_ACCESS_KEY_ID} \
     aws s3 presign \
     --region=${_REGION} \
-    ${S3_URL} )"
+    ${_S3_URL} )"
+}
+
+prepare_ent_urls () {
+  if [[ ${CONSUL_VERSION} == *"ent"* ]]; then
+    export CONSUL_VERSION_STRIPPED=${CONSUL_VERSION/"-ent"/}
+    export CONSUL_ENT_URL=$(presign_ent_url consul ${CONSUL_VERSION_STRIPPED})
+    echo "CONSUL_VERSION_STRIPPED: ${CONSUL_VERSION_STRIPPED}"
+    echo "CONSUL_ENT_URL: ${CONSUL_ENT_URL}"
+    echo "CONSUL_VERSION: ${CONSUL_VERSION}"
+  fi
+
+  if [[ ${VAULT_VERSION} == *"ent"* ]]; then
+    export VAULT_VERSION_STRIPPED=${VAULT_VERSION/"-ent"/}
+    export VAULT_ENT_URL=$(presign_ent_url vault ${VAULT_VERSION_STRIPPED})
+    echo "VAULT_VERSION_STRIPPED: ${VAULT_VERSION_STRIPPED}"
+    echo "VAULT_ENT_URL: ${VAULT_ENT_URL}"
+    echo "VAULT_VERSION: ${VAULT_VERSION}"
+  fi
+
+  if [[ ${NOMAD_VERSION} == *"ent"* ]]; then
+    export NOMAD_VERSION_STRIPPED=${NOMAD_VERSION/"-ent"/}
+    export NOMAD_ENT_URL=$(presign_ent_url nomad ${NOMAD_VERSION_STRIPPED})
+    echo "NOMAD_VERSION_STRIPPED: ${NOMAD_VERSION_STRIPPED}"
+    echo "NOMAD_ENT_URL: ${NOMAD_ENT_URL}"
+    echo "NOMAD_VERSION: ${NOMAD_VERSION}"
+  fi
+}
+
+build () {
+  if [ -z ${RELEASE_VERSION} ]; then
+    # Set RELEASE_VERSION to the current git branch if not specified so it's not empty
+    export RELEASE_VERSION=${GIT_BRANCH}
+  fi
+
+  if [ -z ${USER_TRIGGER+x} ]; then
+    export VCS_NAME="Manual"
+  else
+    export VCS_NAME=${GIT_BRANCH}
+  fi
+
+  echo "Starting build from ${GIT_BRANCH}"
+  echo "RELEASE_VERSION: ${RELEASE_VERSION}"
+  echo "VCS_NAME: ${VCS_NAME}"
+  echo "Building Consul version: ${CONSUL_VERSION}"
+  echo "Building Vault version: ${VAULT_VERSION}"
+  echo "Building Nomad version: ${NOMAD_VERSION}"
+
+  for TEMPLATE in $*; do
+    echo "cd into ${TEMPLATE} directory"
+    cd ${BUILDDIR}/$(echo ${TEMPLATE} | sed 's/-.*//')
+    echo "Building ${TEMPLATE}.json Packer template..."
+
+    if /tmp/packer build ${TEMPLATE}.json ; then
+      echo -e "\033[32m${TEMPLATE} \033[1m[PASS]\033[0m"
+    else
+      echo -e "\033[31m${TEMPLATE} \033[1m[FAIL]\033[0m"
+      return 1
+    fi
+
+    cd -
+  done
+
+  echo "Cleaning up GPG Keyring..."
+  gpg --fingerprint --with-colons ${PGP_SECRET_ID} |\
+    grep "^fpr" |\
+    sed -n 's/^fpr:::::::::\([[:alnum:]]\+\):/\1/p' |\
+    xargs gpg --batch --delete-secret-keys
+
+  echo "Completed build from ${GIT_BRANCH}"
 }
 
 publish () {
+  # Exit early if not on master branch or RUN_PUBLISH env var not passed
+  echo "GIT_BRANCH: ${GIT_BRANCH}"
+  echo "RUN_PUBLISH: ${RUN_PUBLISH}"
+
+  if ! [[ ${GIT_BRANCH} == *"master"* ]] && ! [ ${RUN_PUBLISH} == true ] && ! [ ${RUN_PUBLISH} == 1 ]; then
+    echo "Do not publish, exit early"
+    exit 0
+  fi
+
   export PATH=$PATH:/tmp
   rm -rf tf-se-ami-permissions
   git clone --recurse-submodules https://${GITHUB_USERNAME}:${GITHUB_TOKEN}@github.com/hashicorp/tf-se-ami-permissions
   cd tf-se-ami-permissions
+
   /tmp/terraform init
   ./tfe-cli/bin/tfe pushvars -var "consul_version=${CONSUL_VERSION}" -overwrite consul_version
   ./tfe-cli/bin/tfe pushvars -var "nomad_version=${NOMAD_VERSION}" -overwrite nomad_version
